refactor(tests): extract export_index.rs (1 test, standalone)
Move export_index_uses_named_realm from mod.rs into
engine/tests/export_index.rs.

Open-at-execution decision: standalone file rather than absorbed into
export_docs_index.rs. ExportIndex and ExportDocsIndex are different
CoreOperation variants (DocumentExport vs DocsIndexExport result shapes).
Grouping by operation, not by name similarity.

Tests: 135 passed / 0 failed / 1 ignored (baseline unchanged).

marky-n1h step 8 (4/5).
