+++ markymark-mcp/src/engine/tests/mod.rs
@@ -31,6 +31,7 @@ mod curation;
 mod engine_indexing;
 mod enrich;
 mod export_docs_index;
+mod export_index;
 mod find_references;
 mod from_text_equivalence;
 #[cfg(feature = "semantic-search")]
@@ -42,39 +43,6 @@ mod workspace_scan;
 
 #[cfg(feature = "semantic-search")]
 mod preview_profiling;
-
-#[tokio::test]
-async fn export_index_uses_named_realm() {
-    let dir = make_temp_realm_dir();
-    fs::write(dir.path().join("doc.md"), "# Title\n").unwrap();
-    let engine = make_engine_with_custom_realm("export-realm", dir.path()).await;
-
-    let uri = DocumentUri::from_file_path(&dir.path().join("doc.md"));
-
-    let result = engine
-        .execute(CoreOperation::ExportIndex {
-            uri: uri.clone(),
-            realm: Some("export-realm".to_string()),
-            include_blocks: false,
-        })
-        .await;
-    assert!(
-        matches!(result, CoreOperationResult::DocumentExport { .. }),
-        "expected DocumentExport from named realm, got {result:?}"
-    );
-
-    let result_default = engine
-        .execute(CoreOperation::ExportIndex {
-            uri,
-            realm: None,
-            include_blocks: false,
-        })
-        .await;
-    assert!(
-        matches!(result_default, CoreOperationResult::Error(_)),
-        "expected error from default realm, got {result_default:?}"
-    );
-}
 
 #[tokio::test]
 async fn search_symbols_uses_named_realm() {
